Lower neutron version requirement to queens
This commit, including the reverts before it are made to backport fwaas
ussuri to queens. We reverted all commits which commited import changes
while preserving most of the ussuri refactors and features. This will
hopefully help us to have an easier transition for our to-be-implemented
FWaaS code when we migrate to Ussuri in the future.