On firewall group update, fetch port with admin context
In 329c626 we introduced the capability to bind a firewall group to an
external router port. Turns out, with the user-scoped context you cannot
query for that port, as it is not in the user's scope. Hence
`_validate_ports_for_firewall_group` exits early as it cannot retrieve
the port. The alternative is to query for the port using an admin
context. This is safe as we use it only to retrieve the port and check
the port's attached router against the user's tenant. This would also
fail if the context would not have permission to fetch the router.

Co-authored-by: Sebastian Lohff <sebastian.lohff@sap.com>