Use reader context for DB queries
In most cases FirewallPluginDb has opened reader/writer contexts for its
db access, but not in all.

For us this failed in one particular case, namely when a router port
gets updated and its AFTER_UPDATE methods are called, neutron-fwaas gets
a callback for
neutron_fwaas.services.logapi.common.port_callback.NeutronPortCallBack.handle_event,
where log_db_api.get_logs_for_port() calls
fw_plugin_db.get_fwg_attached_to_port() - outside of a transaction. This
automatically opens a transaction which is never closed, causing a
RuntimeError "Must not be called in a transaction" of the
dhcp_ready_on_ports() provisioning complete callback.

To avoid this, we add a reader context to get_fwg_attached_to_port() and
while we're at it we're adding this to all FirewallPluginDb methods that
do a model_query db query - just to be on the safe side.

Change-Id: Iabb9bc6877059bcc46d64b7fc2d0728e6002610b